+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## unreleased
 
+Changed
+-------
+
+* Skipped ticket ID validation for automatic merge messages.
+* Skipped ticket ID validation in detached HEAD state for rebase and amend commands.
+
+## 0.1.0
+
 Added
 -----